+++ examples/from-scratch/main.go
@@ -13,7 +13,9 @@ func main() {
 
 	vh := doc.AddBlock("VirtualHost", "*:8080")
 	vh.AddDirective("ServerName", "scratch.local")
+	vh.AddComment(" You can change DocumentRoot below")
 	vh.AddDirective("DocumentRoot", "/var/www/scratch")
+	vh.AddComment(" change the /var/www/scratch path", a2cp.WithInlineComment())
 
 	dir := vh.AddBlock("Directory", "/var/www/scratch")
 	dir.AddDirective("Require", "all", "granted")